requirements.txt
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-vector-stores-chroma
huggingface-hub
PyMuPDF
chromadb
ollama
sentence-transformers
llama-index-readers-file
